refactor: change names of current date tasks files
