+++ src/main/factories/usecases/tasks/index.ts
@@ -1,1 +1,1 @@
-export * from './load-current-date-tasks-factory'
+export * from './db-load-current-date-tasks-factory'
